requirements.txt
Flask==1.1.2
torch==1.8.1
Pillow==7.0.0
numpy==1.20.2
requests==2.25.1